+++ lib/cyclone_lariat/errors.rb
@@ -18,5 +18,27 @@ module CycloneLariat
           other.details == details
       end
     end
+
+    class TopicAlreadyExists < LunaPark::Errors::System
+      message { |d| "Topic already exists: `#{d[:expected_topic]}`" }
+    end
+
+    class TopicDoesNotExists < LunaPark::Errors::System
+      message { |d| "Topic does not exists: `#{d[:expected_topic]}`" }
+    end
+
+    class QueueAlreadyExists < LunaPark::Errors::System
+      message { |d| "Queue already exists: `#{d[:expected_queue]}`" }
+    end
+
+    class QueueDoesNotExists < LunaPark::Errors::System
+      message { |d| "Queue does not exists: `#{d[:expected_queue]}`" }
+    end
+    class SubscriptionAlreadyExists < LunaPark::Errors::System
+      message { |d| "Subscription for topic `#{d[:topic].name}`, on endpoint `#{d[:endpoint].name}` already exists" }
+    end
+    class SubscriptionDoesNotExists < LunaPark::Errors::System
+      message { |d| "Subscription for topic `#{d[:topic].name}`, on endpoint `#{d[:endpoint].name}` does not exists" }
+    end
   end
 end