+++ spec/lib/cyclone_lariat_spec.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 require_relative '../../lib/cyclone_lariat/configure'
 
 RSpec.describe CycloneLariat do
@@ -36,11 +38,11 @@ RSpec.describe CycloneLariat do
     end
   end
 
-  describe 'aws_client_id' do
-    subject { cyclone_lariat.aws_client_id }
+  describe 'aws_account_id' do
+    subject { cyclone_lariat.aws_account_id }
 
     context 'when it is defined' do
-      before { cyclone_lariat.aws_client_id = 123 }
+      before { cyclone_lariat.aws_account_id = 123 }
 
       it 'should eq defined value' do
         is_expected.to eq 123