+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+Changed
+- rename `topic` to `queue` for SQS operations, in fact it changed only in methods `publish_event`, `publish_command`
+  if you defined custom queue name
+- rename client_id to account_id
+- send_at no iso8601 format
+- A lot of changes, see README.md
+Added
+- Migrations for create, delete and subscribe topics and queues
+- request_id for Event and Command
+
 ## [0.3.10] - 2022-10-05
 Added
 - Added aws_client_od options
@@ -15,7 +26,6 @@ Added
 - Added configuration options see README.md
 
 ## [0.3.8] - 2022-09-05
-Added
 - Added configuration options see README.md
 
 ## [0.3.8] - 2022-09-05